+++ src/internal.ts
@@ -31,129 +31,146 @@ import { NonEmptyArray } from 'fp-ts/lib/NonEmptyArray'
 // -------------------------------------------------------------------------------------
 
 /** @internal */
-export const isoAsLens = <S, A>(sa: Iso<S, A>): Lens<S, A> => ({
-  get: sa.get,
-  set: flow(sa.reverseGet, constant)
+export const iso = <S, A>(get: Iso<S, A>['get'], reverseGet: Iso<S, A>['reverseGet']): Iso<S, A> => ({
+  get,
+  reverseGet
 })
 
 /** @internal */
-export const isoAsPrism = <S, A>(sa: Iso<S, A>): Prism<S, A> => ({
-  getOption: flow(sa.get, O.some),
-  reverseGet: sa.reverseGet
+export const lens = <S, A>(get: Lens<S, A>['get'], set: Lens<S, A>['set']): Lens<S, A> => ({ get, set })
+
+/** @internal */
+export const prism = <S, A>(
+  getOption: Prism<S, A>['getOption'],
+  reverseGet: Prism<S, A>['reverseGet']
+): Prism<S, A> => ({ getOption, reverseGet })
+
+/** @internal */
+export const optional = <S, A>(getOption: Optional<S, A>['getOption'], set: Optional<S, A>['set']): Optional<S, A> => ({
+  getOption,
+  set
 })
 
 /** @internal */
-export const isoAsOptional = <S, A>(sa: Iso<S, A>): Optional<S, A> => ({
-  getOption: flow(sa.get, O.some),
-  set: flow(sa.reverseGet, constant)
+export const traversal = <S, A>(modifyF: Traversal<S, A>['modifyF']): Traversal<S, A> => ({
+  modifyF
 })
 
 /** @internal */
-export const isoAsTraversal = <S, A>(sa: Iso<S, A>): Traversal<S, A> => ({
-  modifyF: <F>(F: Applicative<F>) => (f: (a: A) => HKT<F, A>) => (s: S) => F.map(f(sa.get(s)), (a) => sa.reverseGet(a))
-})
+export const at = <S, I, A>(at: At<S, I, A>['at']): At<S, I, A> => ({ at })
+
+/** @internal */
+export const index = <S, I, A>(index: Index<S, I, A>['index']): Index<S, I, A> => ({ index })
+
+/** @internal */
+export const isoAsLens = <S, A>(sa: Iso<S, A>): Lens<S, A> => lens(sa.get, flow(sa.reverseGet, constant))
+
+/** @internal */
+export const isoAsPrism = <S, A>(sa: Iso<S, A>): Prism<S, A> => prism(flow(sa.get, O.some), sa.reverseGet)
+
+/** @internal */
+export const isoAsOptional = <S, A>(sa: Iso<S, A>): Optional<S, A> =>
+  optional(flow(sa.get, O.some), flow(sa.reverseGet, constant))
+
+/** @internal */
+export const isoAsTraversal = <S, A>(sa: Iso<S, A>): Traversal<S, A> =>
+  traversal(<F>(F: Applicative<F>) => (f: (a: A) => HKT<F, A>) => (s: S) =>
+    F.map(f(sa.get(s)), (a) => sa.reverseGet(a))
+  )
 
 // -------------------------------------------------------------------------------------
 // Lens
 // -------------------------------------------------------------------------------------
 
 /** @internal */
-export const lensAsOptional = <S, A>(sa: Lens<S, A>): Optional<S, A> => ({
-  getOption: flow(sa.get, O.some),
-  set: sa.set
-})
+export const lensAsOptional = <S, A>(sa: Lens<S, A>): Optional<S, A> => optional(flow(sa.get, O.some), sa.set)
 
 /** @internal */
-export const lensAsTraversal = <S, A>(sa: Lens<S, A>): Traversal<S, A> => ({
-  modifyF: <F>(F: Applicative<F>) => (f: (a: A) => HKT<F, A>) => (s: S) => F.map(f(sa.get(s)), (a) => sa.set(a)(s))
-})
+export const lensAsTraversal = <S, A>(sa: Lens<S, A>): Traversal<S, A> =>
+  traversal(<F>(F: Applicative<F>) => (f: (a: A) => HKT<F, A>) => (s: S) => F.map(f(sa.get(s)), (a) => sa.set(a)(s)))
 
 /** @internal */
-export const lensComposeLens = <A, B>(ab: Lens<A, B>) => <S>(sa: Lens<S, A>): Lens<S, B> => ({
-  get: (s) => ab.get(sa.get(s)),
-  set: (b) => (s) => sa.set(ab.set(b)(sa.get(s)))(s)
-})
+export const lensComposeLens = <A, B>(ab: Lens<A, B>) => <S>(sa: Lens<S, A>): Lens<S, B> =>
+  lens(
+    (s) => ab.get(sa.get(s)),
+    (b) => (s) => sa.set(ab.set(b)(sa.get(s)))(s)
+  )
 
 /** @internal */
-export const prismComposePrism = <A, B>(ab: Prism<A, B>) => <S>(sa: Prism<S, A>): Prism<S, B> => ({
-  getOption: flow(sa.getOption, O.chain(ab.getOption)),
-  reverseGet: flow(ab.reverseGet, sa.reverseGet)
-})
+export const prismComposePrism = <A, B>(ab: Prism<A, B>) => <S>(sa: Prism<S, A>): Prism<S, B> =>
+  prism(flow(sa.getOption, O.chain(ab.getOption)), flow(ab.reverseGet, sa.reverseGet))
 
 /** @internal */
 export const lensComposePrism = <A, B>(ab: Prism<A, B>) => <S>(sa: Lens<S, A>): Optional<S, B> =>
   optionalComposeOptional(prismAsOptional(ab))(lensAsOptional(sa))
 
 /** @internal */
-export const lensId = <S>(): Lens<S, S> => ({
-  get: identity,
-  set: constant
-})
+export const lensId = <S>(): Lens<S, S> => lens(identity, constant)
 
 /** @internal */
-export const lensProp = <A, P extends keyof A>(prop: P) => <S>(lens: Lens<S, A>): Lens<S, A[P]> => ({
-  get: (s) => lens.get(s)[prop],
-  set: (ap) => (s) => {
-    const oa = lens.get(s)
-    if (ap === oa[prop]) {
-      return s
+export const lensProp = <A, P extends keyof A>(prop: P) => <S>(sa: Lens<S, A>): Lens<S, A[P]> =>
+  lens(
+    (s) => sa.get(s)[prop],
+    (ap) => (s) => {
+      const oa = sa.get(s)
+      if (ap === oa[prop]) {
+        return s
+      }
+      return sa.set(Object.assign({}, oa, { [prop]: ap }))(s)
     }
-    return lens.set(Object.assign({}, oa, { [prop]: ap }))(s)
-  }
-})
+  )
 
 /** @internal */
 export const lensProps = <A, P extends keyof A>(...props: readonly [P, P, ...ReadonlyArray<P>]) => <S>(
-  lens: Lens<S, A>
-): Lens<S, { [K in P]: A[K] }> => ({
-  get: (s) => {
-    const a = lens.get(s)
-    const r: { [K in P]?: A[K] } = {}
-    for (const k of props) {
-      r[k] = a[k]
-    }
-    return r as any
-  },
-  set: (a) => (s) => {
-    const oa = lens.get(s)
-    for (const k of props) {
-      if (a[k] !== oa[k]) {
-        return lens.set(Object.assign({}, oa, a))(s)
+  sa: Lens<S, A>
+): Lens<S, { [K in P]: A[K] }> =>
+  lens(
+    (s) => {
+      const a = sa.get(s)
+      const r: { [K in P]?: A[K] } = {}
+      for (const k of props) {
+        r[k] = a[k]
       }
+      return r as any
+    },
+    (a) => (s) => {
+      const oa = sa.get(s)
+      for (const k of props) {
+        if (a[k] !== oa[k]) {
+          return sa.set(Object.assign({}, oa, a))(s)
+        }
+      }
+      return s
     }
-    return s
-  }
-})
+  )
 
 /** @internal */
 export const lensComponent = <A extends ReadonlyArray<unknown>, P extends keyof A>(prop: P) => <S>(
-  lens: Lens<S, A>
-): Lens<S, A[P]> => ({
-  get: (s) => lens.get(s)[prop],
-  set: (ap) => (s) => {
-    const oa = lens.get(s)
-    if (ap === oa[prop]) {
-      return s
+  sa: Lens<S, A>
+): Lens<S, A[P]> =>
+  lens(
+    (s) => sa.get(s)[prop],
+    (ap) => (s) => {
+      const oa = sa.get(s)
+      if (ap === oa[prop]) {
+        return s
+      }
+      const copy: A = oa.slice() as any
+      copy[prop] = ap
+      return sa.set(copy)(s)
     }
-    const copy: A = oa.slice() as any
-    copy[prop] = ap
-    return lens.set(copy)(s)
-  }
-})
+  )
 
 // -------------------------------------------------------------------------------------
 // Prism
 // -------------------------------------------------------------------------------------
 
 /** @internal */
-export const prismAsOptional = <S, A>(sa: Prism<S, A>): Optional<S, A> => ({
-  getOption: sa.getOption,
-  set: (a) => prismSet(a)(sa)
-})
+export const prismAsOptional = <S, A>(sa: Prism<S, A>): Optional<S, A> => optional(sa.getOption, (a) => prismSet(a)(sa))
 
 /** @internal */
-export const prismAsTraversal = <S, A>(sa: Prism<S, A>): Traversal<S, A> => ({
-  modifyF: <F>(F: Applicative<F>) => (f: (a: A) => HKT<F, A>) => (s: S) =>
+export const prismAsTraversal = <S, A>(sa: Prism<S, A>): Traversal<S, A> =>
+  traversal(<F>(F: Applicative<F>) => (f: (a: A) => HKT<F, A>) => (s: S) =>
     pipe(
       sa.getOption(s),
       O.fold(
@@ -161,7 +178,7 @@ export const prismAsTraversal = <S, A>(sa: Prism<S, A>): Traversal<S, A> => ({
         (a) => F.map(f(a), (a) => prismSet(a)(sa)(s))
       )
     )
-})
+  )
 
 /** @internal */
 export const prismModifyOption = <A>(f: (a: A) => A) => <S>(sa: Prism<S, A>) => (s: S): O.Option<S> =>
@@ -191,44 +208,32 @@ export const prismComposeLens = <A, B>(ab: Lens<A, B>) => <S>(sa: Prism<S, A>): 
   optionalComposeOptional(lensAsOptional(ab))(prismAsOptional(sa))
 
 /** @internal */
-export const prismFromNullable = <A>(): Prism<A, NonNullable<A>> => ({
-  getOption: O.fromNullable,
-  reverseGet: identity
-})
+export const prismFromNullable = <A>(): Prism<A, NonNullable<A>> => prism(O.fromNullable, identity)
 
 /** @internal */
-export function prismFromPredicate<A>(predicate: Predicate<A>): Prism<A, A> {
-  return {
-    getOption: O.fromPredicate(predicate),
-    reverseGet: identity
-  }
-}
+export const prismFromPredicate = <A>(predicate: Predicate<A>): Prism<A, A> =>
+  prism(O.fromPredicate(predicate), identity)
 
 /** @internal */
-export const prismSome = <A>(): Prism<O.Option<A>, A> => ({
-  getOption: identity,
-  reverseGet: O.some
-})
+export const prismSome = <A>(): Prism<O.Option<A>, A> => prism(identity, O.some)
 
 /** @internal */
-export const prismRight = <E, A>(): Prism<E.Either<E, A>, A> => ({
-  getOption: O.fromEither,
-  reverseGet: E.right
-})
+export const prismRight = <E, A>(): Prism<E.Either<E, A>, A> => prism(O.fromEither, E.right)
 
 /** @internal */
-export const prismLeft = <E, A>(): Prism<E.Either<E, A>, E> => ({
-  getOption: (s) => (E.isLeft(s) ? O.some(s.left) : O.none), // TODO: replace with E.getLeft in v3
-  reverseGet: E.left
-})
+export const prismLeft = <E, A>(): Prism<E.Either<E, A>, E> =>
+  prism(
+    (s) => (E.isLeft(s) ? O.some(s.left) : O.none), // TODO: replace with E.getLeft in v3
+    E.left
+  )
 
 // -------------------------------------------------------------------------------------
 // Optional
 // -------------------------------------------------------------------------------------
 
 /** @internal */
-export const optionalAsTraversal = <S, A>(sa: Optional<S, A>): Traversal<S, A> => ({
-  modifyF: <F>(F: Applicative<F>) => (f: (a: A) => HKT<F, A>) => (s: S) =>
+export const optionalAsTraversal = <S, A>(sa: Optional<S, A>): Traversal<S, A> =>
+  traversal(<F>(F: Applicative<F>) => (f: (a: A) => HKT<F, A>) => (s: S) =>
     pipe(
       sa.getOption(s),
       O.fold(
@@ -236,7 +241,7 @@ export const optionalAsTraversal = <S, A>(sa: Optional<S, A>): Traversal<S, A> =
         (a) => F.map(f(a), (a: A) => sa.set(a)(s))
       )
     )
-})
+  )
 
 /** @internal */
 export const optionalModifyOption = <A>(f: (a: A) => A) => <S>(optional: Optional<S, A>) => (s: S): O.Option<S> =>
@@ -259,15 +264,12 @@ export const optionalModify = <A>(f: (a: A) => A) => <S>(optional: Optional<S, A
 }
 
 /** @internal */
-export const optionalComposeOptional = <A, B>(ab: Optional<A, B>) => <S>(sa: Optional<S, A>): Optional<S, B> => ({
-  getOption: flow(sa.getOption, O.chain(ab.getOption)),
-  set: (b) => optionalModify(ab.set(b))(sa)
-})
+export const optionalComposeOptional = <A, B>(ab: Optional<A, B>) => <S>(sa: Optional<S, A>): Optional<S, B> =>
+  optional(flow(sa.getOption, O.chain(ab.getOption)), (b) => optionalModify(ab.set(b))(sa))
 
 /** @internal */
-export const findFirst = <A>(predicate: Predicate<A>): Optional<ReadonlyArray<A>, A> => ({
-  getOption: RA.findFirst(predicate),
-  set: (a) => (s) =>
+export const findFirst = <A>(predicate: Predicate<A>): Optional<ReadonlyArray<A>, A> =>
+  optional(RA.findFirst(predicate), (a) => (s) =>
     pipe(
       RA.findIndex(predicate)(s),
       O.fold(
@@ -275,7 +277,7 @@ export const findFirst = <A>(predicate: Predicate<A>): Optional<ReadonlyArray<A>
         (i) => RA.unsafeUpdateAt(i, a, s)
       )
     )
-})
+  )
 
 const unsafeUpdateAt = <A>(i: number, a: A, as: RNEA.ReadonlyNonEmptyArray<A>): RNEA.ReadonlyNonEmptyArray<A> => {
   if (as[i] === a) {
@@ -288,9 +290,8 @@ const unsafeUpdateAt = <A>(i: number, a: A, as: RNEA.ReadonlyNonEmptyArray<A>): 
 }
 
 /** @internal */
-export const findFirstNonEmpty = <A>(predicate: Predicate<A>): Optional<RNEA.ReadonlyNonEmptyArray<A>, A> => ({
-  getOption: RA.findFirst(predicate),
-  set: (a) => (as) =>
+export const findFirstNonEmpty = <A>(predicate: Predicate<A>): Optional<RNEA.ReadonlyNonEmptyArray<A>, A> =>
+  optional<RNEA.ReadonlyNonEmptyArray<A>, A>(RA.findFirst(predicate), (a) => (as) =>
     pipe(
       RA.findIndex(predicate)(as),
       O.fold(
@@ -298,7 +299,7 @@ export const findFirstNonEmpty = <A>(predicate: Predicate<A>): Optional<RNEA.Rea
         (i) => unsafeUpdateAt(i, a, as)
       )
     )
-})
+  )
 
 // -------------------------------------------------------------------------------------
 // Traversal
@@ -306,9 +307,7 @@ export const findFirstNonEmpty = <A>(predicate: Predicate<A>): Optional<RNEA.Rea
 
 /** @internal */
 export function traversalComposeTraversal<A, B>(ab: Traversal<A, B>): <S>(sa: Traversal<S, A>) => Traversal<S, B> {
-  return (sa) => ({
-    modifyF: <F>(F: Applicative<F>) => (f: (a: B) => HKT<F, B>) => sa.modifyF(F)(ab.modifyF(F)(f))
-  })
+  return (sa) => traversal(<F>(F: Applicative<F>) => (f: (a: B) => HKT<F, B>) => sa.modifyF(F)(ab.modifyF(F)(f)))
 }
 
 /** @internal */
@@ -318,12 +317,11 @@ export function fromTraversable<T extends URIS>(T: Traversable1<T>): <A>() => Tr
 export function fromTraversable<T>(T: Traversable<T>): <A>() => Traversal<HKT<T, A>, A>
 /** @internal */
 export function fromTraversable<T>(T: Traversable<T>): <A>() => Traversal<HKT<T, A>, A> {
-  return <A>() => ({
-    modifyF: <F>(F: Applicative<F>) => {
+  return <A>() =>
+    traversal(<F>(F: Applicative<F>) => {
       const traverseF = T.traverse(F)
       return (f: (a: A) => HKT<F, A>) => (s: HKT<T, A>) => traverseF(s, f)
-    }
-  })
+    })
 }
 
 // -------------------------------------------------------------------------------------
@@ -331,39 +329,40 @@ export function fromTraversable<T>(T: Traversable<T>): <A>() => Traversal<HKT<T,
 // -------------------------------------------------------------------------------------
 
 /** @internal */
-export const indexReadonlyNonEmptyArray = <A = never>(): Index<RNEA.ReadonlyNonEmptyArray<A>, number, A> => ({
-  index: (i) => ({
-    getOption: (as) => RA.lookup(i, as),
-    set: (a) => (as) => unsafeUpdateAt(i, a, as)
-  })
-})
+export const indexReadonlyNonEmptyArray = <A = never>(): Index<RNEA.ReadonlyNonEmptyArray<A>, number, A> =>
+  index((i) =>
+    optional(
+      (as) => RA.lookup(i, as),
+      (a) => (as) => unsafeUpdateAt(i, a, as)
+    )
+  )
 
 /** @internal */
-export const indexReadonlyArray = <A = never>(): Index<ReadonlyArray<A>, number, A> => ({
-  index: (i) => ({
-    getOption: (as) => RA.lookup(i, as),
-    set: (a) => (as) =>
-      pipe(
-        RA.updateAt(i, a)(as),
-        O.getOrElse(() => as)
-      )
-  })
-})
+export const indexReadonlyArray = <A = never>(): Index<ReadonlyArray<A>, number, A> =>
+  index((i) =>
+    optional(
+      (as) => RA.lookup(i, as),
+      (a) => (as) =>
+        pipe(
+          RA.updateAt(i, a)(as),
+          O.getOrElse(() => as)
+        )
+    )
+  )
 
 /** @internal */
-export function indexReadonlyRecord<A = never>(): Index<RR.ReadonlyRecord<string, A>, string, A> {
-  return {
-    index: (k) => ({
-      getOption: (r) => RR.lookup(k, r),
-      set: (a) => (r) => {
+export const indexReadonlyRecord = <A = never>(): Index<RR.ReadonlyRecord<string, A>, string, A> =>
+  index((k) =>
+    optional(
+      (r) => RR.lookup(k, r),
+      (a) => (r) => {
         if (r[k] === a || O.isNone(RR.lookup(k, r))) {
           return r
         }
         return RR.insertAt(k, a)(r)
       }
-    })
-  }
-}
+    )
+  )
 
 // -------------------------------------------------------------------------------------
 // At
@@ -371,13 +370,13 @@ export function indexReadonlyRecord<A = never>(): Index<RR.ReadonlyRecord<string
 
 /** @internal */
 export function atReadonlyRecord<A = never>(): At<RR.ReadonlyRecord<string, A>, string, O.Option<A>> {
-  return {
-    at: (key) => ({
-      get: (r) => RR.lookup(key, r),
-      set: O.fold(
+  return at((key) =>
+    lens(
+      (r) => RR.lookup(key, r),
+      O.fold(
         () => RR.deleteAt(key),
         (a) => RR.insertAt(key, a)
       )
-    })
-  }
+    )
+  )
 }